Spin up the ecs agent after user_data has run
If I've got this right, this will hook into the cloud-init configuration
itself. As far as I can tell, runcmd instructions run late in the
instance initialisation stage, so we can configure user_data before we
try to spin up the agent.

See the following pages on the cloud-init documentation:

 * https://cloudinit.readthedocs.io/en/latest/topics/modules.html#runcmd
 * https://cloudinit.readthedocs.io/en/latest/topics/boot.html